requirements.txt
dotted-dict
aiohttp
pytest-aiohttp
requests
